fix and test picklists
Term statuses are different in TBX-Basic from the TBX-Min counterparts,
which are actually equivalent to the display versions in TBX-Basic.
Convert the status picklist to TBX-Basic, and test all values. Also
test the partOfSpeech output values, even though no conversion is
requried.

Output terms and langsets in the order encountered in the input
document. This was required for the new tests to pass.