fix failing command test
There was something weird about passing the name of a program with
quotes around it to `system` that caused it to fail, saying that
`"usr/env/bin/perl"` didn't exist, which it doesn't, technically.
But on Windows it understood that the quotes could be ignored
and on *nix it did not understand that.